Create Draft of inventory functions
This is the main working script for scrambling the base inventory file  in ways that are patterned with randomness features that create relationships between independent and dependent variables. Script uses weighted random functions to "activate" (or not activate) variables. When variables are activated, the script then carries out further random procedures that eventuate to patterned changes in inventories. Randomization factors and sample sizes can be changed to modulate variable relationships. Additional variables can, and need to be, added.