Create Inventory Messup Function
This is the main function to take the base model inventory and mess it up based on what variables are activated. Once a variable is activated (instead of skipped over) various procedures occur involving randomization. Samples of markets, eids, or cids are selected, leading in turn to sample of cids to be manipulated. From the sample of cids, there are three different processes (which of the three occurs is based on weighted probabilities) that lead to manipulation of asset_id's for the particular cid. This changes the distribution of cids with errors in ways that mimic relationships between IVs and DVs