Create Generate new variable for cid age
This is code to add a new set of calculated fields culminating in an integer value of the age of each cid/site. Note that age will be weighed heavily to the year the market was started, reflecting the fact that realistically, a carrier must create a minimum infrastructure and then add to it over the years.